experiment: exclude nonbillable by default
We haven't always tagged "billable" or "nonbillable" and I really don't
want to start doing that, but certain things might need to be excluded
from our billing, and we may need to account for that fully at some
point in the future.

Record-keeping hasn't been diligent enough to show this reliably in the
past - but if it came down to an audit, we could always go back and add
the #nonbillable to things which shouldn't be counted, and run
`hack/since_the_start.sh` to recalculate.

I'm confident that I've been over hours by a substantial margin more
months than I've been under hours, and I'm not genuinely worried about
anything like an audit of my time sheets at this point. These are
personal records and I keep them to make time-keeping easier. But if it
came down to an audit, I would probably change how we count working hours
entirely - move to a "half-day/whole-day" accounting - and we'd have to
count days where I spent 10 minutes working as at least a half-day.

Fortunately we do not have such an adversarial relationship with our
billing department.

Signed-off-by: Kingdon B <kingdon@urmanac.com>